+++ Tests/EasyInjectTests/EasyInjectTests.swift
@@ -8,14 +8,22 @@
 
 import XCTest
 
+protocol LinuxTestCase {
+    associatedtype TestCase: XCTestCase
+    static var allTests: [(String, (TestCase) -> () -> ())] { get }
+}
+
 #if os(Linux)
 public func allTests() -> [XCTestCaseEntry] {
+    let testCases: [LinuxTestCase] = []
     return [
         testCase(StrictInjectorTests.allTests),
         testCase(LazyInjectorTests.allTests),
         testCase(GlobalInjectorTests.allTests),
         testCase(ComposedInjectorTests.allTests),
-        testCase(AnyInjectorTests.allTests)
+        testCase(AnyInjectorTests.allTests),
+        testCase(AnyMutableInjectorTests.allTests),
+        testCase(InjectionErrorTests.allTests)
     ]
 }
 #endif